Move saml function to saml.ts
