+++ server/src/controllers/auth.ts
@@ -23,7 +23,8 @@ import {
   ResetPasswordResponse,
   SystemRole,
 } from '@/common/types';
-import {getSamlStrategy, validateLogin} from './utils/auth';
+import {validateLogin} from './utils/auth';
+import {getSamlStrategy} from './utils/saml';
 import {findAndValidateUserId, findUserById} from './utils/user';
 import {JWT_COOKIE_EXPIRY_MS, JWT_EXPIRY_SECONDS} from '../configs/constants';
 import {JWT_SECRET, NODE_ENV, SAML_SP_CERT_PATH} from '../configs/environment';
